requirements.txt
absl-py==2.1.0
jax==0.5.0
jaxlib==0.5.0
ml_dtypes==0.5.1
numpy==2.2.2
opt_einsum==3.4.0
scipy==1.15.1
